Remove support for multiple sequential sessions
The reason for this is that the reference implementation in Java does
not support this. Therefore it might be confusing if this implementation
would support it. Messages sent to a closed Salt Channel session on Java
disappears into the void. Thus there is not possible to detect if the
Salt Channel session has been closed on the server side from the client